[chore] #190 Organize folder structure in app module
